+++ HPXMLtoOpenStudio/resources/hpxml_defaults.rb
@@ -753,6 +753,7 @@ class HPXMLDefaults
       end
 
       next unless not hvac_control.cooling_setup_temp.nil?
+
       if hvac_control.cooling_setup_start_hour.nil?
         hvac_control.cooling_setup_start_hour = 9 # 9 am
         hvac_control.cooling_setup_start_hour_isdefaulted = true
@@ -1363,6 +1364,7 @@ class HPXMLDefaults
       end
 
       next unless pool.heater_type != HPXML::TypeNone
+
       # Heater
       if pool.heater_load_value.nil?
         default_heater_load_units, default_heater_load_value = MiscLoads.get_pool_heater_default_values(cfa, nbeds, pool.heater_type)
@@ -1416,6 +1418,7 @@ class HPXMLDefaults
       end
 
       next unless hot_tub.heater_type != HPXML::TypeNone
+
       # Heater
       if hot_tub.heater_load_value.nil?
         default_heater_load_units, default_heater_load_value = MiscLoads.get_hot_tub_heater_default_values(cfa, nbeds, hot_tub.heater_type)